requirements.txt
sqlalchemy
pymysql
pandas
lightgbm
flask
pyyaml 